pyproject.toml
[project]
name = "aurora"
version = "0.1.0"
description = "Webhook listener for Evolution API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
python = "^3.8"
